backend/requirements.txt
fastapi~=0.115.0
uvicorn[standard]>=0.30,<0.32
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
tenacity==8.5.0
fitparse==1.2.0
aliyun-python-sdk-core-v3==2.13.33
aliyun-python-sdk-dysmsapi==2.1.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & ORM
SQLAlchemy>=2.0.36,<2.1      # ← 关键：给 3.13 友好的 2.0.x 空间
alembic==1.13.2
pymysql==1.1.0

# Validation & Config
pydantic>=2.7,<2.10          # ← 让 pip 能选到含 3.13 wheel 的 pydantic-core
python-dotenv==1.0.1
